fix evm-substrate-transaction-count
