--- packages/schemas/src/lib/7-evm-substrate-transaction-account/1-0-0.test.ts
@@ -1,18 +1,0 @@
-import * as fs from 'fs';
-
-import { ajv } from '../ajv';
-import { schema } from './1-0-0';
-
-describe('8-evm-substrate-transaction-account/1-0-0', () => {
-  it('should work', () => {
-    const validate = ajv.compile(schema);
-    expect(validate.errors).toBeNull();
-
-    const json = JSON.parse(fs.readFileSync('examples/a8/a8.json', 'utf8'));
-    const valid = validate(json);
-
-    if (!valid) console.log(validate.errors);
-
-    expect(valid).toBeTruthy();
-  });
-});
